Hoàn thiện toàn bộ backend cho các trang trong admin-staff

+++ backend/src/main/java/com/blooddonation/backend/security/jwt/JwtBlacklistService.java
@@ -1,17 +1,15 @@
 package com.blooddonation.backend.security.jwt;
-
 import org.springframework.stereotype.Service;
-
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 @Service
 public class JwtBlacklistService {
     
-    private final ConcurrentMap<String, Long> blacklistedTokens = new ConcurrentHashMap<>();
+    private final ConcurrentMap<String, Integer> blacklistedTokens = new ConcurrentHashMap<>();
     
     public void blacklistToken(String token) {
-        blacklistedTokens.put(token, System.currentTimeMillis());
+        blacklistedTokens.put(token, (int) System.currentTimeMillis());
     }
     
     public boolean isBlacklisted(String token) {
@@ -19,8 +17,8 @@ public class JwtBlacklistService {
     }
     
     public void cleanupExpiredTokens() {
-        long currentTime = System.currentTimeMillis();
+        int currentTime = (int) System.currentTimeMillis();
         blacklistedTokens.entrySet().removeIf(entry -> 
-            currentTime - entry.getValue() > 24 * 60 * 60 * 1000); // 24 hours
+            currentTime - entry.getValue() > 24 * 60 * 60 * 1000);
     }
 }
